fixed overvlow

+++ Common/common.c
@@ -1,10 +1,10 @@
 #include <common.h>
 
 
-double seconds() {
+long double seconds() {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
-	return (double)tv.tv_sec + ((double)tv.tv_usec)/1000000.0;
+	return (long double)tv.tv_sec + ((long double)tv.tv_usec)/1000000.0;
 }
 
 int *allocints(int size) {
